ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
huggingface-hub==0.19.4
transformers==4.36.0
torch==2.1.0
requests==2.31.0
pydantic==2.5.0

